refactor(pandas): split DataFrame into the sunstone.pandas package
Convert sunstone.pandas from a single module into a package and extract
the DataFrame's mixins into focused modules (Spec 0):

- read.py     — ReadMixin (read_csv/read_excel/read_json/read_dataset)
- write.py    — WriteMixin
- metadata.py — MetadataMixin
- ops.py      — OpsMixin
- core.py     — the DataFrame skeleton and _read_tabular_asset helper

sunstone/dataframe.py becomes a thin backwards-compatibility shim that
re-exports DataFrame, _read_tabular_asset, and compute_dataframe_hash.

No behaviour change: the DataFrame MRO is unchanged and PEP 562 lazy-load
is preserved (`import sunstone` pulls in neither pandas nor the package).
